pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "brkraw"
dynamic = ["version"]
description = "Toolkit for loading Bruker Paravision datasets, mapping metadata, and exporting NIfTI"
classifiers = [
    'Development Status :: 5 - Production/Stable',
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "numpy",
    "nibabel",
    "PyYAML",
    "jsonschema",
    "importlib-resources>=5; python_version < '3.9'",
    "typing_extensions",
]
license = {text = "GNLv3"}
maintainers = [
    {name = "SungHo Lee", email = 'shlee@unc.edu'}
]
readme = "README.md"
requires-python = ">=3.8"
keywords = [
    "bruker",
    "paravision",
    "preclinical",
    "mri",
    "neuroimaging",
    "nifti",
    "dataset",
    "conversion",
    "cli",
]

[project.urls]
Homepage = "https://brkraw.github.io"

[project.scripts]
brkraw = "brkraw.cli:main"

[project.entry-points."brkraw.cli"]
addon = "brkraw.cli.commands.addon:register"
hook = "brkraw.cli.commands.hook:register"
config = "brkraw.cli.commands.config:register"
info = "brkraw.cli.commands.info:register"
init = "brkraw.cli.commands.init:register"
session = "brkraw.cli.commands.session:register"
params = "brkraw.cli.commands.params:register"
convert = "brkraw.cli.commands.convert:register"
prune = "brkraw.cli.commands.prune:register"
cache = "brkraw.cli.commands.cache:register"

[project.optional-dependencies]
dev = [
    "mkdocs",
    "mkdocs-material",
    "pymdown-extensions",
    "pytest",
    "tomli",
]
minimal = []
docs = [
    "rich",
]

[tool.hatch.version]
path = "src/brkraw/__init__.py"
style = "pep440"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
artifacts = ["LICENSE"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
